Change Notes: - Moved AudioProjectService.
