+++ Editors/Audio/AudioEditor/AudioProjectData/AudioProjectDataManager.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using Editors.Audio.AudioEditor.AudioProjectData.AudioProjectDataService;
-using Editors.Audio.AudioEditor.AudioProjectData.AudioProjectService;
 using Editors.Audio.AudioEditor.AudioProjectExplorer;
 using Editors.Audio.Storage;
 using Serilog;
