+++ Editors/Audio/AudioEditor/AudioProjectData/AudioProjectDataService/AudioProjectDataServiceParameters.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using Editors.Audio.AudioEditor.AudioProjectData.AudioProjectService;
 using Editors.Audio.Storage;
 
 namespace Editors.Audio.AudioEditor.AudioProjectData.AudioProjectDataService
