+++ Editors/Audio/AudioEditor/AudioProjectExplorer/AudioProjectExplorerViewModel.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using Editors.Audio.AudioEditor.AudioProjectData;
-using Editors.Audio.AudioEditor.AudioProjectData.AudioProjectService;
 using Editors.Audio.Storage;
 using Shared.Core.ToolCreation;
 using Xceed.Wpf.Toolkit;
