+++ Editors/Audio/AudioEditor/NewAudioProject/NewAudioProjectWindow.xaml.cs
@@ -1,5 +1,4 @@
 ﻿using System.Windows;
-using Editors.Audio.AudioEditor.AudioProjectData.AudioProjectService;
 using Shared.Core.PackFiles;
 using Shared.Core.Services;
